% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.R
\name{avgProteinPerCh}
\alias{avgProteinPerCh}
\title{Average proteins per chromosome}
\usage{
avgProteinPerCh(object)
}
\arguments{
\item{object}{a Genome object}
}
\value{
the average number of proteins per chromosome (rounded)
}
\description{
This function calculates the average number of proteins
per chromosome for a Genome object.
}
