% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.R
\docType{class}
\name{Genome-class}
\alias{Genome-class}
\title{Genome class}
\description{
This class stores basic information about a genome,
including its name, number of chromosomes, and number of proteins.
}
\section{Slots}{

\describe{
\item{\code{genome_name}}{the name of the genome}

\item{\code{num_chromosomes}}{number of chromosomes}

\item{\code{num_proteins}}{number of proteins}
}}

